Improves data cleaning
Updated database queries to remove outdated weather data, fixing type issues in time interval calculations within `utils/database.py`.

+++ utils/database.py
@@ -243,7 +243,7 @@ class WeatherDatabase:
                 WHERE location_id = :location_id
                 AND parameter_code = :parameter_code
                 AND forecast_time >= NOW()
-                AND forecast_time <= NOW() + (INTERVAL '1 hour' * :hours_limit::int)
+                AND forecast_time <= NOW() + (INTERVAL '1 hour' * :hours_limit)
                 ORDER BY forecast_time ASC
             """)
             
@@ -463,7 +463,7 @@ class WeatherDatabase:
                 # Delete old forecast data
                 forecast_query = text("""
                     DELETE FROM forecast_data
-                    WHERE created_at < NOW() - (INTERVAL '1 day' * :days::int)
+                    WHERE created_at < NOW() - (INTERVAL '1 day' * :days)
                     OR forecast_time < NOW() - INTERVAL '1 day'
                 """)
                 connection.execute(forecast_query, {"days": days_to_keep})
@@ -472,7 +472,7 @@ class WeatherDatabase:
                 warnings_query = text("""
                     DELETE FROM weather_warnings
                     WHERE (end_time IS NOT NULL AND end_time < NOW())
-                    OR created_at < NOW() - (INTERVAL '1 day' * :days::int)
+                    OR created_at < NOW() - (INTERVAL '1 day' * :days)
                 """)
                 connection.execute(warnings_query, {"days": days_to_keep})
                 
@@ -480,7 +480,7 @@ class WeatherDatabase:
                 model_runs_query = text("""
                     DELETE FROM model_runs
                     WHERE is_latest = FALSE
-                    AND available_at < NOW() - (INTERVAL '1 day' * :days::int)
+                    AND available_at < NOW() - (INTERVAL '1 day' * :days)
                 """)
                 connection.execute(model_runs_query, {"days": days_to_keep})
